+++ amqpstorm/connection.py
@@ -98,6 +98,22 @@ class Connection(Stateful):
         return self._channel0.is_blocked
 
     @property
+    def max_allowed_channels(self):
+        """Returns the maximum allowed channels for the connection.
+
+        :rtype: int
+        """
+        return self._channel0.max_allowed_channels
+
+    @property
+    def max_frame_size(self):
+        """Returns the maximum allowed frame size for the connection.
+
+        :rtype: int
+        """
+        return self._channel0.max_frame_size
+
+    @property
     def server_properties(self):
         """Returns the RabbitMQ Server Properties.
 
@@ -132,6 +148,10 @@ class Connection(Stateful):
 
         with self.lock:
             channel_id = len(self._channels) + 1
+            if channel_id == self.max_allowed_channels:
+                raise AMQPConnectionError(
+                    'reached the maximum number of channels %d' %
+                    self.max_allowed_channels)
             channel = Channel(channel_id, self, rpc_timeout)
             self._channels[channel_id] = channel
             if not lazy: