+++ amqpstorm/tests/unit/channel/channel_message_handling_tests.py
@@ -1,5 +1,6 @@
 import threading
 
+from mock import Mock
 from pamqp import ContentHeader
 from pamqp import specification
 from pamqp.body import ContentBody
@@ -13,7 +14,7 @@ from amqpstorm.tests.utility import TestFramework
 
 class ChannelBuildMessageTests(TestFramework):
     def test_channel_build_message(self):
-        channel = Channel(0, None, 360)
+        channel = Channel(0, Mock(name='Connection'), 360)
 
         message = self.message.encode('utf-8')
         message_len = len(message)
@@ -29,7 +30,7 @@ class ChannelBuildMessageTests(TestFramework):
         self.assertEqual(result.body, message)
 
     def test_channel_build_message_auto_decode(self):
-        channel = Channel(0, None, 360)
+        channel = Channel(0, Mock(name='Connection'), 360)
 
         message = self.message.encode('utf-8')
         message_len = len(message)
@@ -45,7 +46,7 @@ class ChannelBuildMessageTests(TestFramework):
         self.assertEqual(result.body, message.decode('utf-8'))
 
     def test_channel_build_out_of_order_message_deliver(self):
-        channel = Channel(0, None, 360)
+        channel = Channel(0, Mock(name='Connection'), 360)
 
         message = self.message
         message_len = len(message)
@@ -61,7 +62,7 @@ class ChannelBuildMessageTests(TestFramework):
                       self.get_last_log())
 
     def test_channel_build_out_of_order_message_header(self):
-        channel = Channel(0, None, 360)
+        channel = Channel(0, Mock(name='Connection'), 360)
 
         message = self.message
         message_len = len(message)
@@ -78,7 +79,7 @@ class ChannelBuildMessageTests(TestFramework):
                       self.get_last_log())
 
     def test_channel_build_message_headers(self):
-        channel = Channel(0, None, 360)
+        channel = Channel(0, Mock(name='Connection'), 360)
 
         deliver = specification.Basic.Deliver()
         header = ContentHeader(body_size=10)
@@ -91,7 +92,7 @@ class ChannelBuildMessageTests(TestFramework):
         self.assertEqual(result[1].body_size, 10)
 
     def test_channel_build_message_headers_out_of_order(self):
-        channel = Channel(0, None, 360)
+        channel = Channel(0, Mock(name='Connection'), 360)
 
         deliver = specification.Basic.Deliver()
         header = ContentHeader(body_size=10)
@@ -111,7 +112,7 @@ class ChannelBuildMessageTests(TestFramework):
                       self.get_last_log())
 
     def test_channel_build_message_headers_empty(self):
-        channel = Channel(0, None, 360)
+        channel = Channel(0, Mock(name='Connection'), 360)
         channel._inbound = []
         self.assertRaises(IndexError, channel._build_message_headers)
 
@@ -151,7 +152,7 @@ class ChannelBuildMessageTests(TestFramework):
                           channel._build_message_body, 128)
 
     def test_channel_build_message_body(self):
-        channel = Channel(0, None, 360)
+        channel = Channel(0, Mock(name='Connection'), 360)
 
         message = self.message.encode('utf-8')
         message_len = len(message)
@@ -163,7 +164,7 @@ class ChannelBuildMessageTests(TestFramework):
         self.assertEqual(message, result)
 
     def test_channel_build_message_body_break_on_none_value(self):
-        channel = Channel(0, None, 360)
+        channel = Channel(0, Mock(name='Connection'), 360)
 
         message = self.message
         message_len = len(message)
@@ -175,7 +176,7 @@ class ChannelBuildMessageTests(TestFramework):
         self.assertEqual(result, b'')
 
     def test_channel_build_message_body_break_on_empty_value(self):
-        channel = Channel(0, None, 360)
+        channel = Channel(0, Mock(name='Connection'), 360)
 
         message = self.message
         message_len = len(message)